Add JAR File
A JAR File makes it easier and more efficient to deploy and
distribute an entire program, which is why it is added in this version.

Used Gradle to create a JAR File that includes the
classes and resources, along with the JavaFX libraries

Done using Gradle as it is an easy way to do so, and it is the
instruction given to us.