Add assertions to program
The program lacks assertions to ensure various assumptions.

Some important assumptions should hold at various points in the code to
indicate any mistakes in the code.

Let's,
* add assertions to check that the task list loaded is not null.
* add assertions to check tasks in the task list are properply updated.
* add assertions check that the storage written to contains text in the
  proper format.

Using assertions this way can ensure that we have a task list to work
with when the program runs, and verify that the storage contains only
text in the proper format to correctly initialise the task array.