Improve on code quality
Code quality helps to improve readability, which can
save programmers a lot of time when they are debugging.

Remove dead code, structure code logically and reduce
the size of long methods.

There are a few methods that are very long, such as
parse method in Parser class and addTask in TaskList
class. By splitting the codes to more methods, it is
easier to read and update changes.