Remove try-catch block in Parser
IdentifyCommand method in Parser has a try-catch block for cases when a user
does not specify additional required arguments.

There is an inconsistency in some Commands returning their own custom
message if user does not enter additional arguments before a switch
statement was used.

Let's standardize it across the commands for all of them to return
custom messages if user is short of arguments required for command.

It is easier to find out what user argument is lacking from a specified
message rather than a catch-all message.