Add assertion
The code does not have any assertion.

This causes the code to be prone to errors and bad for long-term
maintenance.

Let's add some assert statements.

Most of the assertions are on string as this app is mainly a CLI
based app.

Unnecessarily long commit for such a simple task because this is
a requirement.