Add multiple delete feature User does not have the option to delete multiple objects in one go.
User would manually have to delete each task individually in order to
delete multiple items.

Add multiple delete feature that allows user to input various indexes
separated by a whitespace to delete the tasks at the given positions in
the list.

This would streamline the deleting process for the user.