Add Java assertions
The code mostly depends on a few assertions and if the assertions
failed, Duke will not work. Java Assertions prevent this by reducing
the amount of bugs through alerts while compiling.

Assertions are added at different junctions of the code, mostly the
TaskList as that is where most operations are occurring.

Let's,
* add assertions wherever possible in the future
* not add assertions just to replace exceptions
* not add assertions just to replace JUnit assertions