Add assert statements to code
There are many assumptions made in the execution of this code. In
some cases, if these assumptions fails, it will cause the program
to fail.

Making sure these assumptions are checked improves the code quality.

Let's make sure that these assumptions are checked to avoid any
mistakes by the programmer.