AddCommand: add assert to command flow
If else chain used an else clause, instead of explicitly checking
that the type was "event". This could result in problems if the
type was not one of the three recognised types.

The assumption is now explicit. The code checks sequentially
for each of the three accepted types, and uses an assertion
where the code should not execute.

This checks for a serious error in the execution, since it means a
command was incorrectly parsed as an AddCommand when the syntax was
incorrect.