Improve Code Quality
The code works at the moment; however, the code might be a bit
messy or unclear at specific instances.

The code needs to be refactored as this would improve the
readability, and also simplify various complex expressions
present.

Refactor the code as per the Code Quality guide.

The guide is being followed to refactor to adhere to the coding
standards set up in this module.