Add ability for Duke to detect duplicate tasks when adding tasks
Duke cannot detect duplicate tasks previously, allowing the user to add multiple of
the same tasks. This makes it more user friendly as the user might not remember if
he or she has added a specific task before and thus add it multiple times which can
cause confusion.

Let's override the equals method and define our own equals method in the 3 Task subclasses.

Polymorphism allows us to override the equals method in each of the 3 Task subclasses
to compare the 3 types of tasks differently.  Hence, we can check for any equal task in the
taskList before adding a new task in with just one loop