Duke builds in gradle: extract duke as a jar file using shadowjar in gradle project.
Extracting a jar file allows users to use duke externally in PC without the IDE tool to
open up.

Using johnrengelman's shadowjar plugin, we can easily create a jar file in gradle project. After entering
"gradlew shadowJar", a jar file can be found in /build/libs, the same name as user specifies in shadowJar
config. We do so because it is a powerful feature given in shadowJar.

refer to this link to see more feature about shadowJar.