Enable assertions in gradle run task
Enables Java assertions when running the app.

Use Java assert in parse() method in Parser

Command strings should have matched with the RegEx within the if-else blocks. Assumptions of the string length > n must hold true so that .substring(n) does not throw an Exception. Assumptions of the length of the string array after calling split() must also hold true so that array accesses does not throw an Exception.