Add Assertions
The code works fine. However, there is a risk of it crashing due to
some assumptions that we made during writing the program to become
false during the execution.

Add Assertions at various places to verify that the assumptions made
are true.

Assertions are being adopted because they help define assumptions about
the program state so the runtime can verify them.