Fix bugs from refactor and associated bugs uncovered
Refactor of parse method broke some functions of the application.

Issue was caused by split removing spaces, and join not re-adding
the spaces. In addition, realised that exceptions thrown were not
properly printing error messages to the user.

All the above listed issues were fixed. Errors are now printed
to the user correctly, and when using join, removed spaces are
re-included.