Add assertions to check assumptions
Assumptions when methods are called are never actually ensured.

If there are assumptions that are wrong, then they could cause undefined
behaviour in the code, such as errors.

Let's add assertions to check these assumptions to ensure they are true.

These assumptions are based on what a programmer assumes is true.
Assertions are more suitable compared to exceptions for checking errors
caused by the mistakes of a programmer.