Fix minor bug related to parsing
While the CommandParser can recognise that an input is empty or that a
command does not exist and throw an exception as a result, DukeChatbot
does not handle parsing errors.

If the parsing errors are not handled, then this leads to undefined
behaviour, such as errors like the chatbot not providing feedback to the
user about the fact that empty inputs are not processed.

Let's move the line where the object commandParser parses the input
through its instance method CommandParser::getCommandInstance into the
try block.

Moving the line where the method is called into the try block allows any
exceptions thrown due to parsing invalid inputs to be caught.