Add Java assertions
There are no assertions used in the program.

Assertions document important assumptions that should hold at various
points in the code so that the runtime can verify them.

Add Java assertions at important points in the code, such as during
initialization.

An assertion failure indicates a possible bug in the code. It alerts us
to a possible bug at important points and stops execution to be safe.