Convert exceptions in Command constructor to assertions
The input to the Command constructor is checked if it corresponds to the
Command sub-class. If it does not, an exception is thrown. This check is
also performed in the CommandParser::getCommandInstance method.

This is unnecessary repetition. Also, if the wrong Command sub-class
constructor is called in the CommandParser::getCommandInstance method,
then it must have been the fault of the programmer, because the check
on the user input was already performed in the CommandParser class.

Let's change the exceptions in the Command class to assertions instead.

Exceptions are for errors caused by the user or other external events
such as I/O errors. These remain handled by the CommandParser class. On
the other hand, assertions are for checking for errors caused by
mistakes by the programmer, usually due to wrong assumptions. Thus, the
exceptions in the Command constructor should be converted to assertions
instead.