Add assertions to the project.
The code does not have any assertions to check for bugs in the code.

Using assert feature (not JUnit assertions) to document important
assumptions that should hold at various points in the code can catch
possible bugs.

Assertions are used to define assumptions about the program state so
that the runtime can verify them. An assertion failure indicates a
possible bug in the code because the code has resulted in a program
state that violates an assumption about how the code should behave.