Add assertions to code with message
Assertions help to keep track of logically impossible situations, by
checking if the methods accidentally allow something illogical to bypass it.

Used assert to after some logical statements that should not produce any errors
so as to ensure that the code does not give birth to impossible outcomes.

This is to ensure that the when a soemthing thought to be impossible to produce
actually happens, the program can actually detect it and helps in debugging.