Add JavaFX dependencies to build.gradle file
Gradle can bundle the JavaFX third-party library into the jar file. If
JavaFX is not bundled with the JAR file, the application will not work
in computers that don't have JavaFX installed.

As a step towards preventing this, let's add JavaFX dependencies
to the build.gradle file. Doing so will allow us to bundle the
JavaFX third-party library into the jar file.