Add C-DetectDuplicate
Previous version of duke allows multiple entries of the same task to be
added. It is redundant to add it into our todo-list.

Add method to check if task already exist in the todo-list before adding
it. Gives a using a warning when trying to add duplicated task.

Minor bug fixed in Parser.java: A Deadline task will be added even when
command to add an Event is given.