Improve Code Quality
Code now is messy and sometimes hard to decipher and read.

In order to debug easier in the future, code quality is required to
improve so that others that view the code will also find it more readable.

Improved some lines of code to adhere to guard clauses as well as naming
conventions(Although not all as some of them are a little difficult to change).

Changing naming conventions will allow users to understand the codes and their functions
much easier and guard clauses make the code much more readable.