Storage command: type 'storage' to change storage directory
The current storage directory is fixed to ./data and cannot be changed
by the user.

Allowing the user to change the directory will allow more flexibility
for the user to choose where he/she wants to store their data

I will implement the storage command to open a GUI for which the user
can select their new save directory. The directory will be remembered
even after the application is closed.