Shift parseDate method in Deadline and Event class to Task parent class
Both Deadline and Event class have the same method for parsing a date in
string format.

This breaks the DRY principle as there is a duplication of methods in the
code base.

Let's shift the method to their parent class Task so the method is inherited
instead.

Inheriting the method ensures that there is no duplication of code and the
DRY principle is adhered to.