Storage.java: Added stream and parallelism functionality
The local storage is likely to contain multiple items in more realistic settings.

I have added streams and parallelism functionality to improve the speed of initiating Duke, allowing the user to start using Duke ASAP.

Streams and parallelism allows us to surpass single-core limitations, splitting the work into multiple streams to improve speed of insertion into the ArrayList of things to do.

Please refer to this for more information:
https://nus-cs2103-ay2122s1.github.io/website/schedule/week5/topics.html#w5-7-java-streams