Extract sub-methods in parseCommand for various Command sub-classes
The parseCommand method in the various Command sub-classes tends to be
long. There are many checks and processes that occur within the method
in order to parse the command properly.

Long methods are harder to debug. They also violate software engineering
principles such as SLAP and SRP.

Let's extract various sub-methods from parseCommand for the various
Command sub-classes.

Extracting sub-methods puts a group of lines of code into another
sub-method, replacing those lines with a single line calling that
sub-method, thereby shortening the parseCommand method. The shorter
sub-methods are easier to debug. They also abstract away the
sub-processes done to do checks or to do parsing of the command. The
parseCommand becomes easier to debug, is more readable, and is better at
self-documenting (through the meaningful method names).