Implement tagging of tasks
Tasks can only be searched with keywords associated with its description.

Search function is not very intuitive and useful, since users have to remember the tasks by its description in order to find it in the future.

Let's introduce tagging as a way to classify tasks.

Tagging helps to improve the search function, since users can now search for the category of tasks, rather than a keyword from the description. This will improve usability.