Add BCD-Extension, C-FriendlierSyntax
Currently, code takes in specific commands like delete, todo, deadline, etc.
These commands can be long to type out and thus not feasible.

The commands are long and requires a lot of effort just by typing in.
As such, we need to shorten commands with same functionality in order
to improve user experience.

The code followed C-FriendlierSyntax guidelines and shortened commands in order
for users to have the option to use shortcuts. E.g. deadline can be shortened to
d.

Improves user experience, for example, only needing to type l instead of list to access
their list of tasks.

Welcome message gives a brief rundown of what shortcuts there are, help command also
gives another list of commands that are useful.