Add Assert feature to document important assumptions
No assert features to document important assumptions.

Important assumptions are not being checked and can result in bugs.

Add assert statements to check assumptions. Add assert statements to
ensure the user input passed from function to function is not null.

This prevents failures in the code where user input is necessarily not
null.