Add assertions to the program
There are some places that a certain value or range of value are required.
Using assertions can help to warn users when their is a bug that those values
are not like what we expected.

This change improves the accuracy of the program because we can find bugs
more quickly.

Let's update the code by adding assertions when necessary.