Use assert feature to document important assumptions that should hold at various points
There are some parts in the code that require specific restricted input
from the user. For example, when entering the line number of the task to
delete, the number needs to be a valid number i.e. greater than 0 but
less than or equal to the total number of tasks.

Not validating user input could break code execution or lead to
unexpected outcomes.

Let's add assert statements at the various junctures that require data
validation to catch any invalid user input.

This will help indicate to us the various precautions to take in
anticipation of invalid input from the user.

As far as I can tell, all the potential areas where input validation
could be an issue are either dealt with exceptions or assert statements.