Fix bugs from and finalised archive method
No way to archive methods now, we can only delete
items if we do not want them, and there is no way to save them.

Without archive method, people are not able to store their deleted
tasks, and it may contain important information.

Create an abstraction to manage archiving of information
and deleting from archive when not needed anymore.

We call the class ArchiveCommand which managed the archiving
of tasks and stores them in the ArchiveList class.