Improve tests for Parser and Storage
Unit testing requirements for Parser and Storage are not fulfilled yet. Developer Unit Testing should be done in order verify the function of the Parser and Storage classes.

Improving tests for Parser and Storage will confirm the functionality of both classes and reduce chances of regression.

As a step towards test-driven development, let's build tests in order to test and verify that these classes work.