Add help command
Duke does not show what commands it accepts and what formats users should use to give commands.

It is difficult for user to know what commands Duke accepts.

By adding a "help" command that lists all of Duke's commands and parameters for its commands, users are able to easily see a list of Duke's commands and how to use them.