Add a few important assertions into the code base
We make some assumptions in our code that the task number we give
to `TaskList::getTask` and `TaskList::markTaskAsDone` is within
valid range. We also make some assumptions that the task we want
to mark as done is not null.

This could cause unexpected errors when the tasks we retrieve
ends up being null, or that the task number that we pass to the
function ends up being out of range.

Let's
1. Assert that the task we want to get is not null
2. Assert that the task we want to mark as done is not null
3. Assert that the task number supplied to the
and  is within valid range.