Add assertions in different parts of the code
Assertions are required to check for any possible bugs

Assertions will help us to validate whether the program is
functioning properly by checking our assumptions about
the code at various locations of the code.

Let’s add assertions in suspicious parts of the code where
the result of the code doesn’t turn out the way we expect it to be

Using assertions is preferable in the sense that it might help up to
catch unexpected bugs in our program.