Improve find command to be case insensitive
Find command is case sensitive.

If the user enters a key phrase, the find command only returns
tasks that fit the phrase exactly. (e.g. find book does not return
the task Return Book)

Let's change containsPhrase method in Task to be case insensitive.

This will solve the aforementioned problem of case sensitive keyphrases.

Note: This commit is made on this branch so that I can create a pull
request for branch-A-Streams, since my code already uses streams
and hence was identical to the master branch.