Refactor parser
Code duplication was found when constructing the
flag and the task description from a users input. The parse method which also did the parsing was long.

Divide code into smaller methods to reduce method length. Remove the use of StringBuilder and used Arrays class.

Shorter methods and a KISS approach(by replacing StringBuilder) allow for easier debugging in the future.