Use assertions
Assertions are not used in the code, even though there are assumptions
about the program state which can be defined.

Assertions are used to define assumptions about the program state so
that the runtime can verify them. It can help in identifying possible
bugs in the code.

Add assertions in the relevant command classes and parser class to
define the assumptions of the commands.