Add asserts to Commands
There are several areas in the Commands that require input to be of
a specific format, but no explicit checks are done to confirm this.

This might break the program if nothing is done to check it.

Adding assert statements at critical points of the Command will
ensure that the format is correct.

Assert statements can be disabled and hence will not have any side
effects on the rest of the program.