A-Assertions
Storage functions depending on filePath should not work with an empty filePath.
Task string representation should ensure that the object's status
matches whether it is done.

Storage class asserts that filePath is not empty.
Task class asserts that the valid correct symbol is used.

Assertions in Storage functions can check if the filePath field
is empty.
Assertion in Task toString can check if the object has the correct
status string.