There are assumptions in the code
Making assumptions in the code can lead to possible bugs and the sooner the execution stops, the safer it is.

When the .txt file that the code is trying to read from is missing, we assert that a new directory and file has indeed been created.

Adding these assertions will make sure that the program will stop abruptly when a new directory and file has not been created as they will produce errors when we try to modify the file that does not exist.