Parser class: parses the user’s input and provides the according output
A new command called "undo" helps to reverse the input that was just
made, but its edge cases need to be tweaked.

Let’s check for the previous command is null while simultaneously
checking whether the TaskList is empty.

This is to prevent an exception from being thrown if
"undo" is the user's first input, as a result of the previous input
not yet holding a value.