Add assertions to taskList, Storage, Ui, TaskTime methods
There are several methods that hold certain assumptions that their input will be of the right format. The assertions are there to ensure that if any of those assumption fails, it will be caught

Let's also check if there are any other functions that needs to have an assertion due to the inputs.