Move building of storage string from Storage to TaskList
The buildStorageString method in Storage implicitly understands the
structure of TaskList and Task to be able to build the storage string.

Following OOP principles, TaskList should be able to generate the
storage string on its own and not expose its implementation.

Let's refactor Storage and TaskList and move buildStorageString method
into TaskList as getStorageString.