Parser classes: same function, to decipher the user's input.
The Parser class, with several functions to handle all the commands,
has become too long.

Let’s create another class called Parser functions, to remove out any
functions that do not directly deal with the user input at first. These
functions do not have Command as part of their name. For example,
getItems is only dealt with in Parser’s listCommand function, and can
therefore be relegated to a separate class.

This helps to reduce the size of the Parser class.