Add help command to the chatbot
Users do not have access to the commands usable by the chatbot.

This will cause users to be unable to use the full potential of the
chatbot.

Let's add a help command to the chatbot so that users have access to the
list of commands the chatbot accepts.

This will allow users to use the chatbot more easily.