Refactor code to improve readability
Add Command's methods are too long and it is hard to understand what
is going on.

This might cause Add Command to be hard to maintain as other
developers might not understand it easily.

The original methods are refactored and helper methods are used
to make the code easier to understand.

This is done to allow developers to understand what is going on
easily without needing to understand the underlying implementation.