Allow multiple deletes of tasks in one command
To delete multiple tasks, the user has to manually type the commands one by one for each task, which is repetitive and slow.

Supporting multiple deletes by allowing the user to key in space-separated task numbers to be deleted makes the user experience better.

This could also be easily extended to marking multiple tasks as done in a single command.