Add assertions in Storage.load() and TaskList methods.
* For TaskList methods markTaskDone() and deleteTask(), todoList
should have at least 1 task to work correctly.
* For Storage.load(), the task data parsed into details is
assumed to be formatted correctly containing 3 fields for ToDo
and 4 fields for Deadline and Event.

The assumptions may cause bugs and errors if not met.

Adding assertions will allow assumptions to be verified in runtime
and surface bugs if any.

Using assertions in TaskList is preferable over Exceptions as
the failed validation of taskNum should have already thrown
an Exception in Command before calling the methods in TaskList.