Add Updating of Tasks
Adding class to support extension of IP: Update Tasks.

Previously, updating tasks would require the deletion and re-addition of tasks. A new UpdateCommand class should be created to support the new requirement.

Creating and adding a field to the parser will allow for easy updates of Tasks without having to change anything else. This completes the CRUD functionality.

As a step towards this feature, let's create a new Command, UpdateCommand, and add a case to be handled by the parser.