Add ViewSchedule function
The task list of a user could get very long and it might get hard to
view his immediate tasks. Furthermore, the tasks are not sorted
by date and time.

For a chosen date inputted by the user, the user is able to view all his
tasks on that date. The tasks are sorted in chronological order and
separated by hour. For hours which there are no tasks, the hour
header will not be printed out.

This allows the user to easily see the tasks that needs to be done
for each hour.

The command for this is "schedule dd/mm/yyyy".