Improve code quality
There are some occurrences of magic string literals that appear
in the code. In addition, these string literals appear in numerous
different places.

This would be troublesome when we want to edit
these string literals because we would have to edit multiple
methods. Some examples are the string prefix that we add to all of
Naruto's reply, and the error message given when we encounter an
error while accessing persisted data.

Also, there are multiple unused methods, attributes and
constructors, e.g. the Duke class's default constructor.

These cause unnecessary clutter in the code and make it more
difficult to understand.

Let's
* Save the magic string literals as a final attribute within the
class.
* Remove unused methods, attributes and constructors.