Complete A-Streams
Stream features implemented to improve readability of code

In BotCommandResponderUnit.java, stream is used to add all tasks with the same keyword to a String list for printing.
In BotTaskStatusGeneratorUnit.java, stream is used to convert all tasks in task list to reporting format for reporting.
In BotDynamicMemoryUnit.java. stream is used to convert all tasks in task list to data saving format.