Add assertions to document important assumptions in the code
certain assumptions about the program state are not accounted for

updating the code with assert statements allows these assumptions
to be raised whenever it is met in the program. For example, if the
file path to the text file is empty, raise an Assertion Error.

having assertion statements raise errors indicates that there is a
possible bug and it is safer for the execution of the program stops
this way