Add a way to tag tasks
There is currently no way for the user to add extra notes to a task.

Adding the tag functionality will allow the user to add more information to a task.

As a step towards helping the user categorise the tasks better, let's add a tag functionality so that users can add tags to items.