Implement C-Help
User is unable to know the full range of features that Duke is able to
offer.

Adding a help feature can allow users to check out DUke and its
functionalities.

The entire documentation is stored in a txt file in predetermined file
directory that storage will load and render it on the gui.

It is done this way so that future revision of the features will not
require programmers to edit the code, but instead, just have to edit the
txt file stored in [root]/data/documentation.txt

CHanges to other parts of the code was also being done to accommodate
for the revision in the DukeException class.

.git ignore has also been updated to also include the
data/documentation.txt file in the version control but not the duke.txt
file.