Refactor code to improve code quality
The methods in some classes (specifically those in the utility package)
have the same functionality. Some methods that deal with processing
user input also has convoluted if-else cases. Some input validation was
also done at the wrong place, e.g. Parser checking for duplicate tasks
when TaskList is the one in-charge of adding the task. The inclusion
of many messages in the main body is distracting and causes the main
code body to be cluttered.

The common functionalities across methods leads to code duplication.
Extensive if-else clauses and long methods inhibit readability and
understandability. Having some functionalities in the wrong class might
violate the tell-dont-ask principle.

Let's refactor the code so that everything is in its right place and
clean up the presentation.
* extract duplicate functionalities as a common method
* simplify if-else clauses and use guard clauses wherever necessary so
that code is easier to follow
* shorten long methods by extracting out auxiliary functionalities to
helper methods
* extract messages as constants which are placed away from the main body
of code
* ensure that the tell-dont-ask principle is followed by delegating the
job to the right class (e.g. validation for duplicate tasks should be
done by TaskList)
* other general code quality improvements

Improving code quality and readability makes debugging and further
developing much easier and smooth.