Enhance CodeQuality as per A-CodeQuality
Object fields have been accessed using "this" and does not align with
the Java convention.

This cause certain codes to be unnecessary long.

The this keywords at relevant parts of the code has been removed to
abide by the convention. However, take note that the this keyword has
not been removed from the constructors.

This ensures alignment with convention and readability of code.