Assert task to be not null
Task objects and methods are used in various classes.

If a task object is null, its methods cannot be called and there
will be a nullPointerException.

Therefore task objects are asserted so that they cannot be null.

Assertions will not affect the program but they will notify the
programmer that the task shouldn't be null.