Implement assert feature to document important assumptions
The program state needs to be checked at certain points to ensure
that the code is behaving as expected.

For example, assert is used in the TaskList class, to check whether the
count number that tracks the number of tasks in the tasklist is
properly decremented (when a task is deleted) or incremented (when a
new task is added).

Let's include assertions in all of the methods in TaskList, to ensure
that certain states of the task list are right and error-free.

An AssertionError with details will be thrown if the program state is
not as expected.

Assertions are used to provide more details on what went wrong when
errors are thrown (, that are visible to the author of the code).