Add assertions to Storage class
Empty task may be added to the TaskList while reading from Storage.

Asserting task added in load method to be not null ensures that
no null object is added to TaskList.

Empty task may be written to the storage file.

Asserting task written in writeTasksToFile method to be not empty
or null ensures that no empty strings are written to TaskList and
prevent Null Pointer errors.