Modify Find Date functionality
To implement a schedule viewing function, we can modify our existing
function that finds a list of tasks by date.

We want to arrange the tasks in order of their start time (for events)
or the deadline time (for deadlines). For this we can get Task classes
to implement the Comparable interface, to allow for comparing of
Deadline and Event timings.

Tasks without specified timings can be in a separate list from the timed
ones.

Let's modify Ui and the task package classes to accommodate these new
functionalities.