Add java assertions
There are no checks for important assumptions that should hold at various points in the code.

Lack of check for these key assumptions could lead to possible bugs.

Adding assertions in places where the rest of the program depends on a user defined function
allows us to verify that we have desirable program states throughout.