TaskList class: filter list by using stream
ArrayList tasks is filtered by a for loop, then the filtered task
list will be converted to strings and concatenated to the respond
message in FindTaskCommand and GetListCommand classes.

This violates the tell-don't-ask principle and this process should
be done in the TaskList class.

In the TaskList class, getFilteredListAsString method converts the
task list to a string. The array list is converted to a stream,
filtered, then converted to a string.

In such a way, the command classes can ask for the final string
instead of ask for the list and then convert it to string.