Add help page
The chatbot is relatively easy to use once you get the hang of it.
However, for a new user, it might be a daunting task to find out
more about the commands available, and how to use them.

Add a help page, which shows a list of commands available and
how to use them.

The most intuitive way of looking for help would be by typing help
in the text field.