Refactor to improve the code quality where necessary
Code quality is needed to to guarantee understandability and
readability. This is especially crucial if the code might be passed on
to other developers, or if the code is posted to public platforms such
as Github.

For example, to reduce code duplicate, let's extract the formatting of
messages within each method in the TaskList class, and create a new
method in the Message class to deal with the functionality instead.

Furthermore, messages are extracted to be part of the Message class,
on top of the previously implemented messages.

Some methods may be longer than 30 lines of code, but it is due to the
inclusion of white space and assert.

The unsafe practices mentioned in the Code Quality guide are avoided.