Bye command: make program terminate
Bye command is not terminating the program.

Program still runs and window remains open after executing bye command,
which is not what we expect it to do.

Let's add a Platform.close() call to getResponse. It will only be
called when user gives a Bye command.

This way, the user need not terminate the program by him or herself.