Add a new command called 'update' to Duke
The new command is part of the BCD extension, of which our team chooses category C. It is supposed to update/replace a task at the specified index, and is matched with C-Update.

The command formula is: update + (index number of task to be updated or replaced) + (task type) + (task description).