Add ability to view schedules
The users can view the tasks by using the list command. It allows them
to see which tasks they have completed and which they haven't.

The users need more power and flexibility to be able to view the
tasks they have to do by the end of a particular day.

Add view command that lets the users look at the tasks due on a
particular day.

The ability to specify the date helps the user to customize the
command.