Update GUI for Duke
Components are heavily dependent on each other, and the process of
visually enhancing the GUI is long and painful. In addition, the code is
untidy and long.

Let's refactor the code to use FXML to define our user interface, so
that it is easier to rebuild and control the UI elements in Duke.