Use Assertions
There are several methods that convert a collection to a set. In somerrr
There are a number of assumptions made in the code about the
program's flow of control and input of certain non-public methods.

These assumptions are not checked and could potentially lead to bugs
that are hidden in the program.

Let's use assertions to document these important assumptions that
should hold true during execution of the code.

Assertions do not have a heavy impact on performance and allows
quick detection of bugs arising from the assumptions made in the code.
They also help to document some of the logic in the code and improves
maintainability of the program.