Parser class: Assert for parserDeadline and parserEvent that Deadline and Event contain /by and /at keywords respectively
Deadline and Event need to have /by and /at keywords to separate their task and time of deadline or event.

Let's add an assertion to signal something is wrong with the parse checking since it means an IncorrectInputException is not being returned.

Assertion is being used in addition to the exception check in case there is a missing check for the exception.