Remove magic strings
Using magic strings in the code makes it hard for others to
understand the code.

Let's declare the strings as static constants, with names that
describe what the string is for.

Using a named constant makes the code easier to understand because
the name tells us more about the meaning of the string.