Link Duke's GUI to main program logic
Program logic is not linked to the input entered by the user in Duke's
GUI. The program also does not exit when the user enters 'bye'. This
needs to be updated in order for Duke to becoming fully functional and
have the GUI act as a good representation of the previous CLI.

Let's link the user input in the GUI to Duke's main logic. A new
Enumeration class, DukeStatus, is used to encapsulate the current status
of Duke (e.g. whether there is an error or should exit) as well as the
response that the user should see in the GUI.

Using enumerations is preferable here in order to ensure that only
certain values are allowed, and to encapsulate both the current state of
Duke as well as it's response in a single class.